Fix test suite failure on locales that use comma as decimal mark
The util tests were failing because the string fixtures are using
hard-coded numbers with periods for the decimal marks, but the
function under test was returning localized numbers.

This patch sets the LC_ALL to "C" for the test suite to force all
localized operations to use the standard minimal C locale,
regardless of the user's default system configuration.

For issue: #1