Track processed pacman configuration files and prevent repeats
The mirrorlist file was getting opened and fully processed multiple
times with the default pacman config. It is also possible that
repeated or circular file includes exist for users with complicated
custom pacman configurations.

This patch makes the register_syncs() function track the config
files that it has processed and skip any that it has already seen.