Change `strtrunc_dep_desc()` search to use `strstr()` function
It is possible for a valid optional dependency description to
include colon characters, which could cause the string to be
truncated at the last colon due to the reverse search used in the
`g_strrstr()` function.

This patch changes the colon search to use the `strstr()` function
so that the truncation happens at the first match.