Fix "optional for" dependents lookup using incorrect mechanism
The `alpm_pkg_compute_optionalfor()` function returns package
names, not dependency strings. So looking up the results with a
satisfier search is incorrect behavior.

This could have potentially caused a package providing that name
but sorted higher in the list to be incorrectly shown. However, I
could not find any examples of that scenario to test, so this issue
was probably not causing any user-facing bugs.

This patch changes that lookup to use the `find_package()` function
instead of `find_satisfier()`.