Make `find_satisfier()` prefer installed packages when available
When there are multiple packages providing the dependency, the
result returned by `alpm_find_satisfier()` will simply be the first
package in the list that matches the dependency string requirements.
This could cause the function to return results that are not
relevant to the system being inspected.

This patch alters the satisfier package search to better handle
multiple packages providing the searched dependency by preferring
installed packages, then falling-back to all known packages.

For issue: GH-9